feat(xaa): minimal setup center — servers × people access, advanced behind expand
Collapse the setup center's three sections (people | apps | access) into
two (people | servers). The Access tab is fully absorbed into one
consolidated server list — the people x servers access matrix:

- Row: server name + resource URL, a single Enabled switch (creates the
  connection with scopeMode 'all' on first enable; a re-toggle preserves
  the stored scopeMode/selectedScopes), an assigned-people count whose
  popover assigns/unassigns with a checkbox per active person, and a
  kebab menu (Test in debugger / Edit server / Delete).
- Scope-mode chips, selected-scope editing, per-assignment subsets, and
  auth-server plumbing badges sit behind a collapsed 'Advanced access'
  disclosure per row; the default view never asks about scopes.
- The 'Recent policy decisions' audit footer moves to the bottom of the
  servers section, still collapsed by default.
- Old /xaa-flow/setup/access deep links fall back to the server list;
  the section keeps the 'apps' URL segment.
- XAASetupAccessSection.tsx and its test are deleted; hooks, wire
  contracts, wizard, and the xaa-registration flag gates are untouched.
